karyon_eventemitter::eventemitter

Struct EventListener

pub struct EventListener<T, E> {
    id: EventListenerID,
    recv_chan: Receiver<Event>,
    event_emitter: Weak<EventEmitter<T>>,
    event_id: String,
    topic: T,
    phantom: PhantomData<E>,
}

EventListener listens for and receives events from the EventEmitter.

Fields
id: EventListenerIDrecv_chan: Receiver<Event>event_emitter: Weak<EventEmitter<T>>event_id: Stringtopic: Tphantom: PhantomData<E>
Implementations

impl<T, E> EventListener<T, E>
where T: Hash + Eq + Clone + Debug, E: AsEventValue + Clone,

fn new( id: EventListenerID, event_emitter: Weak<EventEmitter<T>>, recv_chan: Receiver<Event>, event_id: &str, topic: &T, ) -> EventListener<T, E>

Creates a new EventListener.

pub async fn recv(&self) -> Result<E>

Receives the next event from the emitter.

This method blocks until an event is available or the channel is closed. Events are automatically type-cast to the expected type E.

pub fn cancel(&self)

Cancels the event listener and removes it from the EventEmitter.

pub fn topic(&self) -> &T

Returns the topic this listener is registered for.

pub fn event_id(&self) -> &String

Returns the event id for this event listener.

pub fn event_emitter(&self) -> Arc<EventEmitter<T>>

Returns a reference to the parent EventEmitter.